requirements.txt
flask
gunicorn
google-generativeai
textblob
requests